Mark #4694 as fixed (actual fix via jackson-core)
