+++ src/test/java/com/fasterxml/jackson/databind/deser/jdk/JDKNumberDeserTest.java
@@ -18,6 +18,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
 import com.fasterxml.jackson.databind.testutil.DatabindTestUtil;
 
+import static org.junit.Assert.assertEquals;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class JDKNumberDeserTest
@@ -386,4 +387,30 @@ public class JDKNumberDeserTest
         NestedBigDecimalHolder2784 result = mapper.readValue(JSON, NestedBigDecimalHolder2784.class);
         assertEquals(new BigDecimal("5.00"), result.holder.value);
     }
+
+    private final String BIG_DEC_STR;
+    {
+        StringBuilder sb = new StringBuilder("-1234.");
+        // Above 500 chars we get a problem:
+        for (int i = 520; --i >= 0; ) {
+            sb.append('0');
+        }
+        BIG_DEC_STR = sb.toString();
+    }
+    private final BigDecimal BIG_DEC = new BigDecimal(BIG_DEC_STR);
+
+    // [databind#4694]: decoded wrong by jackson-core/FDP for over 500 char numbers
+    @Test
+    public void bigDecimal4694FromString() throws Exception
+    {
+        assertEquals(BIG_DEC, MAPPER.readValue(BIG_DEC_STR, BigDecimal.class));
+    }
+
+    @Test
+    public void bigDecimal4694FromBytes() throws Exception
+    {
+        byte[] b = utf8Bytes(BIG_DEC_STR);
+        assertEquals(BIG_DEC, MAPPER.readValue(b, 0, b.length, BigDecimal.class));
+    }
+
 }
